LayoutTests/platform/ios/accessibility/visible-character-range-height-changes-expected.txt

This tests that visibleCharacterRange returns expected visible ranges with various view height values.


Testing view height values 100 to 1300.
Range with view 500x100, scrollTop 500: {2249, 359}
Range with view 500x200, scrollTop 500: {2249, 809}
Range with view 500x300, scrollTop 500: {2249, 1259}
Range with view 500x400, scrollTop 500: {2249, 1709}
Range with view 500x500, scrollTop 500: {2249, 2159}
Range with view 500x600, scrollTop 500: {2249, 2609}
Range with view 500x700, scrollTop 500: {2249, 3059}
Range with view 500x800, scrollTop 500: {2249, 3489}
Range with view 500x900, scrollTop 500: {1799, 3939}
Range with view 500x1000, scrollTop 500: {1349, 4389}
Range with view 500x1100, scrollTop 500: {899, 4839}
Range with view 500x1200, scrollTop 500: {449, 5289}
Range with view 500x1300, scrollTop 500: {0, 5739}

PASS successfullyParsed is true

TEST COMPLETE
